+++ dpr-calculator/src/main/java/org/devoware/dice/MinusExpression.java
@@ -21,7 +21,7 @@ class MinusExpression implements DieRollExpression {
   }
 
   @Override
-  public int roll() {
+  public double roll() {
     return left.roll() - right.roll();
   }
